% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_heatmap_CE}
\alias{plot_heatmap_CE}
\title{Plot Heatmap of Collateral Effects}
\usage{
plot_heatmap_CE(
  t_result,
  FDR_crit = 0.15,
  species,
  selectedAB = NULL,
  t_or_effect = "effect"
)
}
\arguments{
\item{t_or_effect}{}
}
\value{

}
\description{
Plot Heatmap of Collateral Effects
}
