% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_histogram_CE}
\alias{plot_histogram_CE}
\title{Plot MIC Distribution}
\usage{
plot_histogram_CE(
  MIC_clean,
  results,
  t_rank = 1,
  one_direction = TRUE,
  CResponse = "CS",
  FDR_crit = 0.15,
  whichAB = NULL,
  colours = c("#283c82", "#F46B2D"),
  separate_plots = NULL,
  ran = NULL,
  t_or_effect = "t"
)
}
\arguments{
\item{MIC_clean}{MIC data frame (rows = strains, cols = AB)}

\item{results}{T test results from wrapper function (currently still needed)}

\item{t_rank}{in the \code{results} outcome, which ranking value should be plotted (overwritten by whichAB)}

\item{one_direction}{plot only A on B  (TRUE) or also plot B on A (FALSE)}

\item{CResponse}{direction of response either collateral sensitivity (CS) or resistance (CR)}

\item{FDR_crit}{the criterium for which of results are considered significant}

\item{whichAB}{pair of antibiotics (names string) you want to plot}

\item{colours}{pair or colors for B = high and B = low groups}

\item{separate_plots}{if one_direction = TRUE, whether you want to return two separate plots (in list) or one plot (grid.arrange)}

\item{ran}{range of MIC values plotted}

\item{t_or_effect}{}
}
\value{

}
\description{
Plot MIC Distribution
}
