% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_duplicate_MICs.R
\name{remove_duplicate_MICs}
\alias{remove_duplicate_MICs}
\title{Remove duplicate MIC measurements from MIC data}
\usage{
remove_duplicate_MICs(MIC_df, key = NULL)
}
\arguments{
\item{MIC_df}{a long format data frame with each row an MIC value (\code{MIC}) for
a strain/antibiotic. Contains a \code{measurement_sign} column.
and}

\item{key}{character (vector) containing column names with the columns over
which the duplicates should be removed.}
}
\value{
data frame with same structure as \code{MIC_df}, with unique rows based on
the provided key
}
\description{
If multiple MIC measurements are done on the same strain and antibiotic, but
this function will summarize MIC value based on \code{measurement_sign}
}
\examples{
#not yet created
}
