% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_antibiotic.R
\name{translate_antibiotic}
\alias{translate_antibiotic}
\title{Translate antibiotic names to abbreviations}
\usage{
translate_antibiotic(
  antibiotic_names,
  dictionary,
  which_dictionary_cols = c("antibiotic", "abbreviation")
)
}
\arguments{
\item{antibiotic_names}{character vector containing the names to be
translated.}

\item{dictionary}{matrix or data frame with at least two columns containing
the names to be translated and the names to which they will be translated.}

\item{which_dictionary_cols}{character vector with two column names of the
from and to translation values, respectively.}
}
\value{
character vector with same length as \code{antibiotic_names} containing
translated names
}
\description{
Translate antibiotic names to abbreviations
}
\examples{
dict <- data.frame(antibiotic = c("amikacin", "tobramycin", "colistin"),
                   abbreviation = c("AMK", "TOB", "CST"))
ab_names <- c("amikacin", "amikacin", "tobramycin")
translate_antibiotic(ab_names, dict)
}
