% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collateral_mult_test.R
\name{collateral_mult_test}
\alias{collateral_mult_test}
\title{Test Multiple Collateral Effects}
\usage{
collateral_mult_test(
  MIC_data,
  effect_type = "both",
  crit_type = "median",
  criterium = NULL,
  MTC_method = "BY"
)
}
\arguments{
\item{MIC_data}{data frame or matrix with MIC for strains (rows) and
antibiotics (cols, with corresponding colnames)}

\item{effect_type}{type of collateral effect to be evaluated, string:
\code{"both"}, \code{"CR"} or \code{"CS"}}

\item{crit_type}{type of dichotomization criterion, string: \code{"quant"},
\code{"median"} or \code{"log2_MIC"}}

\item{criterium}{value of dichotomization criterium, numeric value (is
overwritten by \code{crit_type = "median"})}
}
\value{
data frame with results per antibiotic combination
}
\description{
Wrapper for \code{collateral_t_test()}, which quantifies and tests collateral
effects in a set of MICs for more than two antibiotics.
}
