% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_histogram_CE}
\alias{plot_histogram_CE}
\title{Plot MIC Distribution}
\usage{
plot_histogram_CE(
  A,
  B,
  effect_type = "both",
  crit_type = "median",
  criterium = NULL,
  MIC_range = NULL,
  colors = c("#BFC6B8", "#4A5242")
)
}
\arguments{
\item{A}{Untransformed MIC values of antibiotic A, numeric vector}

\item{B}{Untransformed MIC values of antibiotic B, numeric vector, same length as A}

\item{effect_type}{type of collateral effect to be evaluated, string:
\code{"both"}, \code{"CR"} or \code{"CS"}}

\item{crit_type}{type of dichotomization criterion, string: \code{"quant"},
\code{"median"} or \code{"log2_MIC"}}

\item{criterium}{value of dichotomization criterium, numeric value (is
overwritten by \code{crit_type = "median"})}

\item{MIC_range}{range of x-axis (default NULL)}

\item{colors}{set of two colors for the high and low MIC groups}
}
\value{
ggplot object with histogram
}
\description{
Plot the distribution of MIC of A in a histogram, split on B.
}
\examples{
MIC_test <- data.frame(d1 = c(16, 16, 128, 128, 32, 16, 128, 1, 64, 1),
                       d2 = c(32, 32, 16, 32, 8, 32, 8, 8, 8, 8),
                       d3 = c(8, 64, 64, 64, 16, 6, 16, 8, 8, 8))
MIC_test_result <- collateral_mult_test(MIC_test)
plot_histogram_CE(MIC_test$d1, MIC_test$d2)


}
